requirements.txt
Django==4.0.2
djangorestframework==3.13.1
pandas
psycopg2==2.9.3
PyYAML